move the metadata
